+++ src/containers/Form/CustomMetrics/index.js
@@ -1,25 +1,18 @@
 /*
  * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
  */
- 
+
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { reduxForm } from 'redux-form/immutable';
-import CustomMetricsForm from './Layout';
+import CustomMetricsForm from './layout';
 import { onFormSubmit } from '../../../redux/modules/entities';
 import { formValidation } from './validation';
-import {
-  getSelectedEntityId,
-  isInherited,
-  isCreating
-} from '../../../redux/modules/entities/selectors';
+import { getSelectedEntityId, isInherited, isCreating } from '../../../redux/modules/entities/selectors';
 import { selectFormInitialValues } from '../../../redux/modules/form/selectors';
-import {
-  getAbandonTypeFormValue
-} from '../../../redux/modules/entities/customMetrics/selectors';
+import { getAbandonTypeFormValue } from '../../../redux/modules/entities/customMetrics/selectors';
 
-export const formSubmission = (values, dispatch, props) =>
-  dispatch(onFormSubmit(values, props));
+export const formSubmission = (values, dispatch, props) => dispatch(onFormSubmit(values, props));
 export const createFormName = state => ({
   form: `customMetrics:${getSelectedEntityId(state)}`
 });