+++ src/containers/Form/EmailTemplates/__tests__/layout.test.js
@@ -5,13 +5,13 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import EmailTemplatesForm from '../Layout';
+import EmailTemplatesForm from '../layout';
 
 describe('<EmailTemplatesForm />', () => {
   let customTemplates;
   beforeEach(() => {
     customTemplates = [
-      "Hello {{{user-first-name}}} {{{user-last-name}}}:↵↵You've been invited to following tenant on {{{product-name}}}: {{{tenant-name}}}↵↵Simply log in with your existing {{{product-name}}} username, {{{username}}}.↵↵Please click the following link to get started on accepting this invitation using single sign on: {{{sso-config-login-url}}}↵↵Note that the link above will expire after 24 hours. If you have forgotten your password or have issues please contact your system administrator for assistance.↵↵Thank you,The Serenova Team!",
+      "Hello {{{user-first-name}}} {{{user-last-name}}}:↵↵You've been invited to following tenant on {{{product-name}}}: {{{tenant-name}}}↵↵Simply log in with your existing {{{product-name}}} username, {{{username}}}.↵↵Please click the following link to get started on accepting this invitation using single sign on: {{{sso-config-login-url}}}↵↵Note that the link above will expire after 24 hours. If you have forgotten your password or have issues please contact your system administrator for assistance.↵↵Thank you,The Serenova Team!"
     ];
   });
   it('renders default email', () => {